Fix BlendMode enum: cast GL_NONE/GL_SRC_ALPHA to GLint for glbinding compatibility
imageoverlaygl.h:70 - BlendMode enum has GLint underlying type but GL_NONE and
GL_SRC_ALPHA are GLenum (enum class) in glbinding, requiring explicit cast.

Agent-Logs-Url: https://github.com/inviwo/inviwo/sessions/cd57c4bb-34b3-4732-bda3-04fa02e6f88a

+++ modules/basegl/include/modules/basegl/processors/imageprocessing/imageoverlaygl.h
@@ -67,7 +67,10 @@ class Outport;
 class IVW_MODULE_BASEGL_API OverlayProperty : public CompositeProperty {
 public:
     enum class Positioning { Relative, Absolute };
-    enum class BlendMode : GLint { Replace = GL_NONE, Over = GL_SRC_ALPHA };
+    enum class BlendMode : GLint {
+        Replace = static_cast<GLint>(GL_NONE),
+        Over = static_cast<GLint>(GL_SRC_ALPHA)
+    };
 
     OverlayProperty(std::string identifier, std::string displayName,
                     InvalidationLevel invalidationLevel = InvalidationLevel::InvalidOutput,
